// src/lib/types.ts
export interface Book {
    id: number;
    title: string;
    author: string;
    country: string;
    language: string;
    link: string;
    pages: string;
    year: string;
  }
  
  export interface BookFormData {
    title: string;
    author: string;
    country: string;
    language: string;
    link: string;
    pages: string;
    year: string;
  }